Updates + Code clean up
added Building,Trees,Crops and Stage Managers to clean up core script. we also have stage#.js files now which is used to load different scenes. it houses the core map loading and position of buildings, trees, and crops for the current scene.

also fixed a bug with 2 players entering and exiting buildings causing it to completely vanish due to a race condition.